Refactor map.js to display pins from FirestoreThe diff shows changes to the map.js file, where the code has been refactored to display pins retrieved from Firestore on a map. The getPins() function has been imported from the firestore.js file, and the forEach loop has been replaced with a promise chain to retrieve pins asynchronously. The code now sets up markers on the map for each pin and binds popups with the desired information
